Split legislature tests out to their own file
The test class is far too big, and we're about to add even more to it.
So separate out the legislature tests first.
